Refactor phone formatation
