+++ app/src/utils/removeSpecialChars.js
